backend/database.py
import mysql.connector

def connect_db():
    conn = mysql.connector.connect(
        host="localhost",
        user="root",
        password="root75",
        database="library_db"
    )
    return conn