requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
Werkzeug==3.1.5
